# Propagate iteration to parent in OffsetArrays