This provides a performance boost, especially for ranges but also to some extent for arrays.

On master
```julia
julia> r = 1:1:1000; ro = OffsetArray(r); s = collect(r); so = OffsetArray(s);

julia> v = ones(1000);

julia> @btime $v[$r];
  2.010 μs (1 allocation: 7.94 KiB)

julia> @btime $v[$ro];
  25.848 μs (1 allocation: 7.94 KiB)

julia> @btime $v[$s];
  1.897 μs (1 allocation: 7.94 KiB)

julia> @btime $v[$so];
  3.859 μs (1 allocation: 7.94 KiB)
```

After this PR
```julia
julia> @btime $v[$ro];
  1.679 μs (1 allocation: 7.94 KiB)

julia> @btime $v[$so];
  1.862 μs (1 allocation: 7.94 KiB)
```