@timholy 2021-08-19
It really would be nice not to have to specialize so many things on OffsetArray. The more methods we have, the less generic we are.  https://github.com/JuliaArrays/OffsetArrays.jl/pull/255#issuecomment-901784749 seems like it needs a decision before we do things like this.