@jishnub 2021-08-19
This is unrelated to bounds-checking though, and may be addressed independent of #255. This seems to be a necessary addition at least for ranges, and may be version-limited if `Base` decides to use traits to check if the argument is range-like for iteration. I can restrict this to `OffsetRange`s to avoid making the method too broad, if this is a concern.