@timholy 2021-08-19
My point is that it's better to address these by private methods or traits; we don't want to overload every exported function in Base specifically for OffsetArrays, that would be a nightmare. As much as possible, we should give them the generic interface and then let the generic implementations handle them.

I recognize that may not always be possible, but that should be the goal.