@timholy 2021-08-19
Another approach: what is the specific source of the poor performance on master? Can we fix it there instead?