@jishnub 2021-08-19
That might be possible using a trait-based approach similar to #255. Perhaps the source of this difference, at least for `OffsetRange`s, is that `iterate` doesn't realize that the argument is range-like, and supports rapid iteration using `first`, `step` and `last` without a need to index into the argument. The difference for arrays is a bit harder to track down, perhaps more efficient bounds-checking?